+++ gogas-service/frontend/src/components/ActionDialog.tsx
@@ -8,7 +8,16 @@ import {
   Button
 } from '@material-ui/core';
 
-const ActionDialog = ({title, message, open, onCancel, onAction, actions}) => {
+interface Props {
+  title: string;
+  message: string;
+  open: boolean;
+  onCancel: () => void;
+  onAction: (actionIndex: number) => void;
+  actions: string[];
+}
+
+const ActionDialog: React.FC<Props> = ({ title, message, open, onCancel, onAction, actions }) => {
   return (
     <Dialog open={open} onClose={onCancel} maxWidth='xs' fullWidth>
       <DialogTitle>{title}</DialogTitle>
@@ -16,11 +25,11 @@ const ActionDialog = ({title, message, open, onCancel, onAction, actions}) => {
       <DialogContent>{message}</DialogContent>
 
       <DialogActions>
-      <Button onClick={onCancel} autoFocus>
+        <Button onClick={onCancel} autoFocus>
           Annulla
         </Button>
-        {actions.map((action,i) => (
-          <Button key={`action-${i}`}  onClick={() => onAction(i)}>
+        {actions.map((action, i) => (
+          <Button key={`action-${i}`} onClick={() => onAction(i)}>
             {action}
           </Button>
         ))}
