+++ gogas-service/frontend/src/components/LoadingListItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ListItem, CircularProgress } from '@material-ui/core';
 
-const LoadingListItem = () => {
+const LoadingListItem: React.FC = () => {
   return (
     <ListItem disableGutters style={{ justifyContent: 'center' }}>
       <CircularProgress />
