+++ gogas-service/frontend/src/components/navigationmenu/types.d.ts
@@ -1,0 +1,12 @@
+export interface MenuChapter {
+  label?: string;
+  items: MenuItem[];
+}
+
+export interface MenuItem {
+  label: string;
+  url: string;
+  icon: number;
+  restrictions?: string[];
+}
+
